Impromptu Backend Changes
Modified framework to be more efficient for local context requests.
Also published a new set of context providers.

IMPORTANT:  There have been some slight changes to how information is
passed to GCF Android and Desktop applications.  See the template
projects to see the changes (most of which are very slight).